fix: fix movies card
