+++ src/components/App/App.js
@@ -23,6 +23,7 @@ function App() {
         <Routes>
           <Route path='/' element={<Main />} />
           <Route path='/movies' element={<Movies />} />
+          <Route path='/saved-movies' element={<Movies />} />
           {/* <Route path='/profile' element={<Profile />} />
           <Route path='/sign-up' element={<Registration />} />
           <Route path='/sign-in' element={<Login />} />
