+++ src/components/MoviesCard.jsx/MoviesCardDeleteBtn/MoviesCardDeleteBtn.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import './MoviesCardDeleteBtn.scss';
+
+export default function MoviesCardDeleteBtn() {
+  return (
+    <button
+      button="button"
+      className="movies__delete"
+    >
+    </button>
+  )
+}
